Add infrastructure to override example responses/requests in ReDoc docs. Fix airmass and max_allowable_ipp.
Due to limitations in DRF's serializers, we are unable to create serializers that describe some of our API endpoints. Add ability to specify example requests/responses and override what is generated by DRF. Any view that implements get_example_request or get_example_response and returns a dictionary example response will have that response used as an example in the docs.